requirements.txt
openai-whisper>=20231117
yt-dlp>=2024.4.9
torch>=2.0.0
tqdm>=4.66.0


# Testing
pytest>=7.0.0
